+++ notebooks/hosp_only_ww_model.qmd
@@ -1,5 +1,5 @@
 ---
-title: "Replicating Hospital Only Model from `cdcgov/wastewater-informed-covid-forecasting`"
+title: "Replicating Hospital Only Model from ww-inference-model"
 format: gfm
 engine: jupyter
 ---
@@ -31,9 +31,10 @@ This script also saves the posterior samples from the model for comparison to ou
 
 ## Load Data and Create Priors
 
-We begin by loading the stan_data, converting it the correct inputs for our model, and definitng the model.
+We begin by loading the Stan data, converting it the correct inputs for our model, and definitng the model.
 
 ```{python}
+# | label: create model
 my_hosp_only_ww_model, data_observed_hospital_admissions = (
     create_hosp_only_ww_model_from_stan_data(
         "data/fit_hosp_only/stan_data.json"
@@ -80,6 +81,7 @@ posterior_predictive = my_hosp_only_ww_model.posterior_predictive(
 ## Prepare for plotting
 
 ```{python}
+# | label: prepare for plotting
 import arviz as az
 
 idata = az.from_numpyro(
@@ -92,16 +94,19 @@ idata = az.from_numpyro(
 ## Plot Predictive Distributions
 
 ```{python}
+# | label: plot prior preditive
 plotting.plot_predictive(idata, prior=True)
 ```
 
 ```{python}
+# | label: plot posterior preditive
 plotting.plot_predictive(idata)
 ```
 
 ## Plot all posteriors
 
 ```{python}
+# | label: plot all posteriors
 for key in list(idata.posterior.keys()):
     try:
         plotting.plot_posterior(idata, key)
